requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Flask==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
Werkzeug==2.3.6
